requirements.txt
# Meeting Bot Dependencies (Simplified)

# Web server
fastapi>=0.100.0
uvicorn>=0.20.0

# HTTP client
httpx>=0.25.0

# Configuration management
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Browser automation for meeting join
playwright>=1.40.0

# Date/time utilities
python-dateutil>=2.8.0

# AWS SDK for S3 uploads
boto3>=1.34.0
